+++ .github/workflows/wheel_macos_x86_64.yml
@@ -26,7 +26,7 @@ env:
 
   CIBW_BUILD_VERBOSITY: 1
 
-  MACOSX_DEPLOYMENT_TARGET: '10.14'
+  MACOSX_DEPLOYMENT_TARGET: 10.14
 
 jobs:
   mac-wheels-x86:
