+++ pennylane_lightning/src/simulator/DynamicDispatcher.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 /**
- * @file
- * Defines
+ * @file DynamicDispatcher.hpp
+ * Defines DynamicDispatcher class. Can be used to call a gate operation by string.
  */
 
 #pragma once
